auto-improve: [#799 Step 4/6] Extract small/isolated `cmd_*` functions from `cai.py` → `cai_lib/cmd_misc.py`

Refs damien-robotsix/robotsix-cai#821
